fix warnings

+++ src/MacMain.m
@@ -323,9 +323,9 @@ static void CustomApplicationMain (int argc, char **argv)
 
 - (NSString *)stringByReplacingRange:(NSRange)aRange with:(NSString *)aString
 {
-    unsigned int bufferSize;
-    unsigned int selfLen = [self length];
-    unsigned int aStringLen = [aString length];
+    unsigned long bufferSize;
+    unsigned long selfLen = [self length];
+    unsigned long aStringLen = [aString length];
     unichar *buffer;
     NSRange localRange;
     NSString *result;
